+++ src/main/java/frc/robot/commands/GalacticSearch/PathChooserCommandGroupB.java
@@ -42,7 +42,7 @@ public class PathChooserCommandGroupB extends InstantCommand {
     // CameraData cameraData = m_camera.createData();
 
     if (m_camera.getRegions() != null) {
-      int positionY = m_camera.findClosestRegion(m_camera.getRegions()).getTopBound();
+      double positionY = m_camera.findClosestRegion(m_camera.getRegions()).getTopBound();
       Logger.Log("PositionY:", 1, "" + positionY);
       if (positionY > 140) {
         Logger.Log("Path Chooser B", 1, "Choosing close path");
